@@ -11,7 +11,7 @@
         }
     ],
     "require": {
-        "php": "^7.4 || ^8.0",
+        "php": "^8.0",
         "ext-simplexml": "*",
         "ext-json": "*",
         "psr/log": "^1.0 || ^2.0 || ^3.0",
@@ -26,11 +26,6 @@
     "autoload": {
         "psr-4": {
             "AlibabaCloud\\Oss\\V2\\": "src"
-        }
-    },
-    "autoload-dev": {
-        "psr-4": {
-            "AlibabaCloud\\Oss\\V2\\Tests": "tests"
         }
     }
 }
