@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace AlibabaCloud\Oss\V2\Signer;
 
-class NopSigner implements SignerInterface
+class SignerV4 implements SignerInterface
 {
     public function sign(SigningContext $signingCtx)
     {
-        return null;
+
     }
 }
